\name{barplotSignatures-methods}
\docType{methods}
\alias{barplotSignatures-methods}
\alias{barplotSignatures,epheno,character-method}
\alias{barplotSignatures,epheno,GeneSetCollection-method}
\alias{barplotSignatures,epheno,GeneSet-method}
\alias{barplotSignatures,epheno,list-method}
\title{Methods for Function barplotSignatures in Package `phenoTest'}
\description{
Methods for function \code{barplotSignatures} in Package
`phenoTest'. For more information read the function's manual.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "epheno", signatures = "character")}}{
Method for an \code{epheno} object and one signature stored in an object
of class \code{character}.
}

\item{\code{signature(x = "epheno", signatures = "GeneSetCollection")}}{  
Method for an \code{epheno} object and several signatures stored in an
object of class \code{GeneSetCollection}.
}

\item{\code{signature(x = "epheno", signatures = "GeneSet")}}{        
Method for an \code{epheno} object and one signature stored in an object
of class \code{GeneSet}.
}

\item{\code{signature(x = "epheno", signatures = "list")}}{
Method for an \code{epheno} object and several signatures stored in an
object of class \code{list}.
}

}}
\keyword{methods}
