\name{barplotSignifSignatures-methods}
\docType{methods}
\alias{barplotSignifSignatures-methods}
\alias{barplotSignifSignatures,epheno,character-method}
\alias{barplotSignifSignatures,epheno,list-method}
\alias{barplotSignifSignatures,epheno,GeneSet-method}
\alias{barplotSignifSignatures,epheno,GeneSetCollection-method}
\title{Methods for Function barplotSignifSignatures in Package `phenoTest'}
\description{
Methods for function \code{barplotSignifSignatures} in Package
`phenoTest'. For more information read the function's manual.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "epheno", signatures = "character")}}{
Method for an \code{epheno} object and one signature stored in an object
of class \code{character}.
}

\item{\code{signature(x = "epheno", signatures = "list")}}{
Method for an \code{epheno} object and several signatures stored in an
object of class \code{list}.
}

\item{\code{signature(x = "epheno", signatures = "GeneSet")}}{
Method for an \code{epheno} object and one signature stored in an object
of class \code{GeneSet}.
}

\item{\code{signature(x = "epheno", signatures = "GeneSetCollection")}}{
Method for an \code{epheno} object and several signatures stored in an
object of class \code{GeneSetCollection}.
}
}}
\keyword{methods}
