\name{export2CSV-methods}
\docType{methods}
\alias{export2CSV-methods}
\alias{export2CSV,epheno-method}
\title{ Methods for Function export2CSV in Package 'phenoTest'}
\description{
 Methods for function \code{export2CSV} in Package 'phenoTest'
}
\section{Methods}{
\describe{
\item{\code{signature(x = "epheno")}}{ Exports summary differences
  (fold changes, hazard ratios), p-values and gene annotation (when
  available) to a CSV (comma separated value) file }
}}
\keyword{methods}

