\name{getVars2test-methods}
\docType{methods}
\alias{getVars2test-methods}
\alias{getVars2test,epheno-method}
\title{Methods for Function getVars2test in Package `phenoTest'}
\description{
Methods for function \code{getVars2test} in Package `phenoTest'. For
more information read the funcion's manual.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "epheno")}}{
Method for an object of class \code{epheno}.
}
}}
\keyword{methods}
