\name{gsea.selGsets}
\alias{gsea.selGsets}
\alias{gsea.selVars}
\title{
Subset an object of class gseaData.
}
\description{
\code{gsea.selGsets} to select a subgroup of gene sets from an object of
class gseaData.
\code{gsea.selVars} to select a subgroup of variables from an object of
class gseaData.
}
\usage{
gsea.selGsets(x, selGsets)
gsea.selVars(x, selVars)

}
\arguments{
  \item{x}{ an object of class \code{gseaData}.}
  \item{selGsets}{ names of the gene sets that we want to keep.}
  \item{selVars}{ names of the variables that we want to keep.}
}
\value{
Returns an object of class \code{gseaData}.
}
\author{
Evarist Planet
}
\keyword{ ~kwd1 }

