\name{heatmapPhenoTest-methods}
\docType{methods}
\alias{heatmapPhenoTest-methods}
\alias{heatmapPhenoTest,ExpressionSet,character-method}
\alias{heatmapPhenoTest,ExpressionSet,list-method}
\alias{heatmapPhenoTest,ExpressionSet,missing-method}
\alias{heatmapPhenoTest,ExpressionSet,GeneSet-method}
\alias{heatmapPhenoTest,ExpressionSet,GeneSetCollection-method}
\title{ Methods for Function heatmapPhenoTest in Package `phenoTest'}
\description{
Methods for function \code{heatmapPhenoTest} in Package `phenoTest'. For
more information read the function's manual.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "ExpressionSet", signatures = "character")}}{
Method for an \code{ExpressionSet} object and one signature stored in an
object of class \code{character}.
}

\item{\code{signature(x = "ExpressionSet", signatures = "list")}}{
Method for an \code{ExpressionSet} object and several signatures stored
in an object of class \code{list}.
}

\item{\code{signature(x = "ExpressionSet", signatures = "missing")}}{
Method for an \code{ExpressionSet} object and no signatures.
}

\item{\code{signature(x = "ExpressionSet", signatures = "GeneSet")}}{
Method for an \code{ExpressionSet} object and one signature stored in an
object of class \code{GeneSet}.
}

\item{\code{signature(x = "ExpressionSet", signatures = "GeneSetCollection")}}{
Method for an \code{ExpressionSet} object and several signatures stored
in an object of class \code{GeneSetCollection}.
}
}}
\keyword{methods}