\name{htmlpage}
\alias{htmlpage}
\docType{data}
\title{
htmlpage
}
\description{
This is a copy of the \code{htmlpage} function from annotate
\code{package}. The copy was made to allow htmlpage to produce links to
local files.
}
\details{
See \code{?annotate::htmlpage} fror more information.
}
\keyword{datasets}
