\name{pAdjust-methods}
\docType{methods}
\alias{pAdjust-methods}
\alias{pAdjust,epheno-method}
\title{ Methods for Function pAdjust in Package `phenoTest'}
\description{
Methods for function \code{pAdjust} in Package `phenoTest'. This
function adjusts the p-values of an \code{epheno} object. For more
informe read the function's manual.
}
\section{Methods}{
\describe{

\item{\code{signature(x = "epheno")}}{
Adjusts the pvalues of an epheno object.
}
}}
\author{
Evarist Planet
}
\keyword{methods}
