\name{phenoTest-package}
\alias{phenoTest-package}
\alias{phenoTest}
\docType{package}
\title{
Test correlation between gene expression and phenotype.
}
\description{
Test correlation between gene expression and phenotype.
}
\details{
\tabular{ll}{
Package: \tab phenoTest\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-04-28\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Evarist Planet
Maintainer: Evarist Planet <evarist.planet@irbbarcelona.org>
}